chore(overview_cache): Move overview handling to repository/overview_cache, write unit tests for overview cache
